Fixes issue with linker AdditionalDependencies including unexpected "%3B" characters, instead of a semi-collon. The solution was using spaces instead of semicolons when setting AdditionalDependencies.
Related issues:
#openframeworks/visualstudioPlugin/issues/1
