+++ src/Wizard/Wizard.csproj
@@ -65,6 +65,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Addon.cs" />
+    <Compile Include="Extensions.cs" />
+    <Compile Include="Folder.cs" />
     <Compile Include="FormOFRoot.cs">
       <SubType>Form</SubType>
     </Compile>
