+++ src/source.extension.vsixmanifest
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <PackageManifest Version="2.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2011" xmlns:d="http://schemas.microsoft.com/developer/vsx-schema-design/2011">
     <Metadata>
-        <Identity Id="VSIXopenframeworks2017.A049B6B4-2855-4C1C-A3A6-13ED823BB898" Version="0.5" Language="en-US" Publisher="Arturo Castro, Half Scheidl" />
+        <Identity Id="VSIXopenframeworks2017.A049B6B4-2855-4C1C-A3A6-13ED823BB898" Version="0.6" Language="en-US" Publisher="Arturo Castro, Half Scheidl" />
         <DisplayName>openFrameworks plugin for Visual Studio 2017</DisplayName>
         <Description xml:space="preserve">openFrameworks plugin for Visual Studio 2017</Description>
         <MoreInfo>https://openframeworks.cc/</MoreInfo>
